requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dtw-python==1.3.0
entrypoints==0.4
fastapi==0.93.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gradio==3.20.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0
linkify-it-py==2.0.0
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai-whisper==20230307
orjson==3.8.7
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
pooch==1.7.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
resampy==0.4.2
rfc3986==1.5.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.4
starlette==0.25.0
threadpoolctl==3.1.0
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
yarl==1.8.2
zipp==3.15.0
